# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MSM Alimentos — API pública",
    "version": "1.2.0",
    "description": "Endpoints públicos do site institucional MSM. Para integradores (MullerBot, monitoramento, automações). Responses incluem header `X-API-Version` (`1.2`). CORS aberto em endpoints públicos (`*`); endpoints com auth Bearer não devem ser chamados de browser sem origin allowlist explícita.",
    "contact": { "name": "MSM Alimentos", "email": "comercial@msm.com.br" }
  },
  "servers": [
    { "url": "https://site-msm-production.up.railway.app", "description": "Produção (temporário Railway)" },
    { "url": "http://localhost:3000", "description": "Local dev" }
  ],
...
  "paths": {
...
    "/api/forms/submit": {
      "post": {
        "tags": ["public"],
        "summary": "Submeter formulário de contato",
        "description": "Pipeline: honeypot → Turnstile → rate limit → Zod → MullerBot. Sempre responde 200 mesmo se MullerBot falhar (entra em retry queue).",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": { "type": "string", "minLength": 8, "maxLength": 128, "pattern": "^[A-Za-z0-9_\\-:.]+$" },
            "description": "Chave opcional para deduplicar retries. Cache 24h."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/FormSubmitInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sucesso (ou honeypot disparado silenciosamente)",
            "headers": {
              "X-Idempotent-Replay": {
                "schema": { "type": "string", "enum": ["true"] },
                "description": "Presente apenas quando a resposta veio do cache de idempotency"
              }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          },
          "400": {
            "description": "Validação falhou (Zod, Turnstile, JSON, ou Idempotency-Key inválida)",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "429": {
            "description": "Rate limit excedido",
            "headers": {
              "Retry-After": { "schema": { "type": "integer" }, "description": "Segundos até reset" },
              "X-RateLimit-Remaining": { "schema": { "type": "integer" } },
              "X-RateLimit-Reset": { "schema": { "type": "integer" }, "description": "Unix timestamp" }
            }
          },
          "500": { "description": "Falha ao enfileirar (Supabase down)" }
        }
      }
    },
    "/api/lgpd/consent": {
      "post": {
        "tags": ["public"],
        "summary": "Registrar consentimento LGPD",
        "description": "Envia hash sha256 + version do texto LGPD ao MullerBot (fire-and-forget). Não armazena PII.",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": { "type": "string", "minLength": 8, "maxLength": 128, "pattern": "^[A-Za-z0-9_\\-:.]+$" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["accepted"],
                "properties": {
                  "accepted": { "type": "boolean", "description": "true = aceitou cookies, false = recusou" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Consentimento registrado" },
          "400": { "description": "Body inválido" }
        }
      }
    },
    "/api/revalidate": {
      "get": {
        "tags": ["internal"],
        "summary": "Revalidação on-demand de ISR",
        "parameters": [
          { "name": "path", "in": "query", "schema": { "type": "string" }, "description": "Path a invalidar (ex: /produtos)" },
          { "name": "tag", "in": "query", "schema": { "type": "string" }, "description": "Tag de cache a invalidar (ex: footer)" }
        ],
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Revalidado" },
          "401": { "description": "REVALIDATE_SECRET inválido" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Header `Authorization: Bearer <SECRET>` (REVALIDATE_SECRET ou CRON_SECRET conforme rota)"
      }
    },
    "schemas": {
...
      "FormSubmitInput": {
        "type": "object",
        "required": ["form_type", "interest_type", "name", "email", "whatsapp", "lgpd_consent"],
        "properties": {
          "form_type": {
            "type": "string",
            "enum": ["representante", "food_service", "b2b", "terceirizacao", "envase", "contato_geral"]
          },
          "interest_type": {
            "type": "string",
            "enum": ["food_service", "b2b", "terceirizacao", "envase", "marcas_proprias", "distribuicao", "representante"]
          },
          "name": { "type": "string", "minLength": 2, "maxLength": 120 },
          "email": { "type": "string", "format": "email", "maxLength": 120 },
          "whatsapp": { "type": "string", "description": "Aceita formato BR com ou sem máscara; normalizado para +55..." },
          "lgpd_consent": { "type": "boolean", "enum": [true], "description": "Obrigatório true" },
          "message": { "type": "string", "maxLength": 2000 },
          "source_page": { "type": "string", "maxLength": 200, "default": "/contato" },
          "company": { "type": "string", "maxLength": 160 },
          "city": { "type": "string", "maxLength": 80 },
          "state": { "type": "string", "maxLength": 2, "description": "UF (será uppercased)" },
          "segment": { "type": "string", "maxLength": 160 },
          "estimated_volume": { "type": "string", "maxLength": 160 },
          "operation_type": { "type": "string", "maxLength": 160 },
          "product_interest": { "type": "string", "maxLength": 160 },
          "product_type": { "type": "string", "maxLength": 160 },
          "packaging_type": { "type": "string", "maxLength": 160 },
          "region": { "type": "string", "maxLength": 120 },
          "experience": { "type": "string", "maxLength": 500 },
          "captcha_token": { "type": "string", "description": "Token Cloudflare Turnstile" },
          "website": {
            "type": "string",
            "maxLength": 0,
            "description": "HONEYPOT — deve permanecer vazio. Submissões com este campo preenchido são silenciosamente descartadas."
          }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "required": ["ok"],
        "properties": {
          "ok": { "type": "boolean", "enum": [true] },
          "message": { "type": "string" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["ok"],
        "properties": {
          "ok": { "type": "boolean", "enum": [false] },
          "message": { "type": "string" },
          "error": { "type": "string" }
        }
      }
    }
  }
}